Changed AnswerMapper to use MapStruct Moved Csv parsing methods to CSVParser
